requirements.txt
streamlit
python-dotenv
openai
requests
pytest
pytest-mock